+++ packages/storage/src/typed.rs
@@ -71,7 +71,7 @@ where
         let mapped = self
             .storage
             .range(start, end, order)?
-            .map(|item| item.and_then(deserialize_kv::<T>));
+            .map(deserialize_kv::<T>);
         Ok(Box::new(mapped))
     }
 
@@ -133,7 +133,7 @@ where
         let mapped = self
             .storage
             .range(start, end, order)?
-            .map(|item| item.and_then(deserialize_kv::<T>));
+            .map(deserialize_kv::<T>);
         Ok(Box::new(mapped))
     }
 }